# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Courier map API for St. Petersburg & Leningrad Oblast
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: pois
  - name: routes
  - name: geo
  - name: stack
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stack/status:
    get:
      operationId: getStackStatus
      tags: [stack]
      summary: Status of self-hosted geo services (Martin, GraphHopper, Pelias)
      responses:
        "200":
          description: Status report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackStatus"

  /pois:
    get:
      operationId: listPois
      tags: [pois]
      summary: List POIs (optionally filtered by bbox and types)
      parameters:
        - name: bbox
          in: query
          description: "minLng,minLat,maxLng,maxLat"
          schema: { type: string }
        - name: types
          in: query
          description: Comma-separated POI types
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 1000, maximum: 5000 }
      responses:
        "200":
          description: POI list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poi"
    post:
      operationId: createPoi
      tags: [pois]
      summary: Create a personal POI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePoiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poi"

  /pois/recent:
    get:
      operationId: listRecentPois
      tags: [pois]
      summary: Recently created POIs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Recent POIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poi"

  /pois/stats:
    get:
      operationId: getPoiStats
      tags: [pois]
      summary: Counts grouped by POI type
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoiStats"

  /pois/h3:
    get:
      operationId: getPoiHexAggregation
      tags: [pois]
      summary: H3 hex-grid aggregation of POIs
      parameters:
        - name: resolution
          in: query
          schema: { type: integer, default: 8, minimum: 5, maximum: 11 }
      responses:
        "200":
          description: Hex aggregation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/H3Bucket"

  /pois/nearby:
    get:
      operationId: listNearbyPois
      tags: [pois]
      summary: POIs within radius of a point
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lng
          in: query
          required: true
          schema: { type: number }
        - name: radius
          in: query
          required: true
          description: Radius in meters
          schema: { type: number, default: 500 }
      responses:
        "200":
          description: Nearby POIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoiWithDistance"

  /pois/{id}:
    get:
      operationId: getPoi
      tags: [pois]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: POI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poi"
        "404":
          description: Not found
    patch:
      operationId: updatePoi
      tags: [pois]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePoiInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poi"
    delete:
      operationId: deletePoi
      tags: [pois]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  /routes:
    get:
      operationId: listCourierRoutes
      tags: [routes]
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourierRoute"
    post:
      operationId: createCourierRoute
      tags: [routes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourierRouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierRoute"

  /routes/{id}:
    delete:
      operationId: deleteCourierRoute
      tags: [routes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  /geo/geocode:
    get:
      operationId: searchAddress
      tags: [geo]
      summary: Forward geocoding (via Pelias when up, falls back to Nominatim)
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: Geocoding results
          content:
            application/json:
              schema:
                type: object
                required: [results, source]
                properties:
                  source: { type: string }
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/GeocodeResult"

  /geo/reverse:
    get:
      operationId: lookupCoordinates
      tags: [geo]
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lng
          in: query
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Reverse geocoding result
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source: { type: string }
                  label:
                    type: ["string", "null"]
                  address:
                    type: ["object", "null"]
                    additionalProperties: { type: string }

  /geo/route:
    get:
      operationId: routeBetween
      tags: [geo]
      summary: Route between two points (GraphHopper e-bike profile when up, OSRM fallback)
      parameters:
        - name: from
          in: query
          required: true
          description: "lat,lng"
          schema: { type: string }
        - name: to
          in: query
          required: true
          description: "lat,lng"
          schema: { type: string }
        - name: profile
          in: query
          schema:
            type: string
            enum: [ebike, bike, foot, car]
            default: ebike
      responses:
        "200":
          description: Routing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    StackStatus:
      type: object
      required: [martin, graphhopper, pelias, postgis, basemap]
      properties:
        martin:
          $ref: "#/components/schemas/ServiceHealth"
        graphhopper:
          $ref: "#/components/schemas/ServiceHealth"
        pelias:
          $ref: "#/components/schemas/ServiceHealth"
        postgis:
          $ref: "#/components/schemas/ServiceHealth"
        basemap:
          type: object
          required: [source, url]
          properties:
            source: { type: string, enum: [pmtiles_martin, raster_osm_fallback] }
            url: { type: string }

    ServiceHealth:
      type: object
      required: [up, label]
      properties:
        up: { type: boolean }
        label: { type: string }
        detail:
          type: ["string", "null"]

    Poi:
      type: object
      required: [id, type, title, lng, lat, createdAt, updatedAt]
      properties:
        id: { type: integer }
        type:
          type: string
          enum: [entrance, gate, passage, note, parking, stairs]
        title: { type: string }
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        lng: { type: number }
        lat: { type: number }
        h3R9:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    PoiWithDistance:
      allOf:
        - $ref: "#/components/schemas/Poi"
        - type: object
          required: [distanceM]
          properties:
            distanceM: { type: number }

    CreatePoiInput:
      type: object
      required: [type, title, lng, lat]
      properties:
        type:
          type: string
          enum: [entrance, gate, passage, note, parking, stairs]
        title: { type: string, minLength: 1 }
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        lng: { type: number }
        lat: { type: number }

    UpdatePoiInput:
      type: object
      properties:
        type:
          type: string
          enum: [entrance, gate, passage, note, parking, stairs]
        title: { type: string, minLength: 1 }
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        lng: { type: number }
        lat: { type: number }

    PoiStats:
      type: object
      required: [total, byType]
      properties:
        total: { type: integer }
        byType:
          type: array
          items:
            type: object
            required: [type, count]
            properties:
              type: { type: string }
              count: { type: integer }

    H3Bucket:
      type: object
      required: [h3, count, lng, lat]
      properties:
        h3: { type: string }
        count: { type: integer }
        lng: { type: number }
        lat: { type: number }

    CourierRoute:
      type: object
      required: [id, name, coordinates, distanceM, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description:
          type: ["string", "null"]
        coordinates:
          type: array
          description: GeoJSON LineString coordinates [lng, lat]
          items:
            type: array
            items: { type: number }
        distanceM:
          type: ["number", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    CreateCourierRouteInput:
      type: object
      required: [name, coordinates]
      properties:
        name: { type: string, minLength: 1 }
        description:
          type: ["string", "null"]
        coordinates:
          type: array
          minItems: 2
          items:
            type: array
            minItems: 2
            maxItems: 2
            items: { type: number }

    GeocodeResult:
      type: object
      required: [label, lng, lat]
      properties:
        label: { type: string }
        lng: { type: number }
        lat: { type: number }
        source: { type: string }

    GeocodeResponse:
      type: object
      required: [results, source]
      properties:
        source: { type: string }
        results:
          type: array
          items:
            $ref: "#/components/schemas/GeocodeResult"

    ReverseGeocodeResponse:
      type: object
      required: [source]
      properties:
        source: { type: string }
        label:
          type: ["string", "null"]
        address:
          type: ["object", "null"]
          additionalProperties: { type: string }

    RouteStep:
      type: object
      required: [text, distanceM, durationS]
      properties:
        text: { type: string }
        distanceM: { type: number }
        durationS: { type: number }

    RouteResponse:
      type: object
      required: [profile, source, distanceM, durationS, coordinates, steps]
      properties:
        profile: { type: string }
        source: { type: string }
        distanceM: { type: number }
        durationS: { type: number }
        coordinates:
          type: array
          items:
            type: array
            items: { type: number }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RouteStep"