# lib/api-zod/src/generated/types/courierRoute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courier map API for St. Petersburg & Leningrad Oblast
 * OpenAPI spec version: 0.1.0
 */

export interface CourierRoute {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** GeoJSON LineString coordinates [lng, lat] */
  coordinates: number[][];
  /** @nullable */
  distanceM: number | null;
  createdAt: Date;
  updatedAt: Date;
}