# lib/api-zod/src/generated/types/createCourierRouteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courier map API for St. Petersburg & Leningrad Oblast
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCourierRouteInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @minItems 2 */
  coordinates: number[][];
}