# lib/api-zod/src/generated/types/routeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courier map API for St. Petersburg & Leningrad Oblast
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStep } from "./routeStep";

export interface RouteResponse {
  profile: string;
  source: string;
  distanceM: number;
  durationS: number;
  coordinates: number[][];
  steps: RouteStep[];
}