# lib/api-zod/src/generated/types/routeStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Courier map API for St. Petersburg & Leningrad Oblast
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStep {
  text: string;
  distanceM: number;
  durationS: number;
}